chore: extract wallet client methods in plain js functions
